backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
aiogram==3.13.1
python-dotenv==1.0.0
aiohttp==3.10.5
